canonicalize tml files
We put the sections in this canonical order:

- in-yaml
- in-json
- error
- out-yaml
- emit-yaml
- test-event

We also run all the in-json sections through jq and standardise the field
ordering (lexically by key).

There should be no semantic changes in this PR. The only manual intervention was to FRK4.tml, which has a comment in the in-json section. By my reading of the point_lines production in http://testml.org/specification/language/index.html (which I *think* is in effect here), it seems like the comment is literal (included in the JSON data), not a TML-level comment, so that would make it invalid JSON. I've left the comment in anyway for now, and manually applied jq to the actual JSON.